move adapter to new file
