+++ languages/js/src/index.ts
@@ -11,5 +11,6 @@ export {
   FilterCondition,
   Adapter,
 } from './filter';
+export { typeOrmAdapter } from './typeorm_adapter';
 export { defaultEqualityFn } from './helpers';
 export type { Class } from './types';
